Rename questions.js to polls.js
